comment/requirements.txt
azure-core==1.26.4
azure-functions
azure-cosmos==4.5.1
